Support for custom install dirs in CMake build -- 5
This restores support for custom install directories, which was lost
when moving from autotools to CMake. This is needed in particular by
distributions to build versioned LLVM packages which are
parallel-installable with the latest version.

This introduces new CMAKE_INSTALL_* variables, following the naming
convention used by the GNUInstallDirs CMake module: for BINDIR, LIBDIR,
INCLUDEDIR, DOCDIR and MANDIR. Existing variables are preserved, and
take their defaults from or determine the new ones' defaults as
appropriate.