Allow for CMake modules to not be installed in lib/cmake/llvm
A common use-case of GNUInstallDirs to separate run time and
"development time" items. In this case, the libs are run time
(especially when dynamic linking), but the CMake modules are development
time. We therefore add another variable to customize where it goes
analogous to the GUNInstallDirs ones.

It is called LLVM_INSTALL_CMAKE_DIR because the other packages use that
(among other variables) to hold the output --cmakedir from llvm-config.
We are configuring that, so it seemed appropriate to use the same one.