Remove `COMPILER_RT_INSTALL_PATH`
It turns out we don't neeed to remove `clang/runtime` to make this work,
as @phosek thought might be acceptable is in fact not.

To do this, we just set `CMAKE_INSTALL_PREFIX` rather than
`COMPILER_RT_INSTALL_PATH` in `clang/runtime` in the child CMake
invocation. Since it's a separate CMake invocation, we don't have to
worry about mutation "bleeding over" and effecting other projects.

However I am still worrie about redefining of `CMAKE_INSTALL_PREIX` in
`compiler-rt/cmake/base-config-ix.cmake`. This occurs under
`LLVM_TREE_AVAILABLE`. If this is just for separate compiler-rt builds
(separate invocation of CMake, that is), this is fine. But if on the
other hand this is for combination builds in a single CMake run, then
this sort of redefinition is no good as it will effect more than
compiler-rt.

Differential Revision: https://reviews.llvm.org/D99860