Remove `clang/runtime`
This patch removes `clang/runtime`, after `COMPILER_RT_INSTALL_PATH` is
removed in order to simplify things for D99484.

As @phosek said in https://reviews.llvm.org/D99484#2662806

> I'm not sure if we still need `COMPILER_RT_INSTALL_PATH`. That
> variable is only used by `clang/runtime/CMakeLists.txt` which
> predates `runtimes/CMakeLists.txt` and was AFAIK only ever used by
> Apple. I think  we should consider removing
> `COMPILER_RT_INSTALL_PATH`. We'll need to check if
> `clang/runtime/CMakeLists.txt` is still being used or not.

It turns out `COMPILER_RT_INSTALL_PATH` isn't only used by that, but we
might (hopefully!) be able to remove it annyways. If so, `clang/runtime`
can still be removed, which is a simplification we might as well do if
@phosek is right that it is no longer in use.

Differential Revision: https://reviews.llvm.org/D99755