Make `llvm_install_symlink` robust with respect to absolute dirs.
If `CMAKE_INSTALL_BINDIR` is a different absolute path per project, as
it is with NixOS when we install every package to its own prefix, the
old way fails when the downstream packages try to install symlinks in
the now-immutable `LLVM_TOOLS_INSTALL_DIR`.

The new way ensures the downstream projects' prefixes will be used
instead, as desired.

Differential Revision: https://reviews.llvm.org/D101070