[Clang][LLD][cmake] Drop deprecated support for `llvm-config`-based build
Summary:
This has been deprecated for Clang since D51714 (2018), and LLD since
D116492 (earlier in 2022).

That means both methods have been deprecated for at least two releases
(14 and 15), and Clang may more.

Not sure what the deprecation policy is, but I am opening this now for
whenever it's OK to land.

Reviewers: sebastian-ne	mstorsjo compnerd beanz phosek

Subscribers: